03-demos/vogue-concierge/pyproject.toml
# Copyright 2026 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "vogue-concierge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "colorama>=0.4.6",
    "google-adk[extensions]>=1.29.0",
    "honcho>=2.0.0",
    "pillow>=12.2.0",
    "python-dotenv>=1.0.1",
    "requests>=2.33.1",
]

[tool.uv.workspace]
members = [
    "vogue_concierge_app/backend",
]
